aadd

+++ TeamProject/src/team/control/MainController.java
@@ -51,10 +51,13 @@ public class MainController implements Initializable {
 	@FXML
 	TextField txtFieldTitle;
 
+	private Diary updateDiary;
 	private User userInfo;
-	private Stage primaryStage;
 	private TableView<Diary> tableViewUserList;
 	private boolean stop;
+	private int tableViewPreNextNum = 1;
+	private int userListCount = 0;
+	private int listPageNum = 1;
 
 	public void setUserInfo(User userInfo) {
 		this.userInfo = userInfo;
@@ -66,9 +69,17 @@ public class MainController implements Initializable {
 			@Override
 			public void run() {
 				changeDate();
+				ConnectionDAO cDAO = new ConnectionDAO();
+				userListCount = cDAO.getUserListCount(userInfo.getUserid());
 				btnMyList.setOnAction(e -> userList(e));
-				hyLinkLogout.setOnMouseClicked(event -> primaryStage.close());
+				hyLinkLogout.setOnMouseClicked(event -> {
+					((Stage) hyLinkLogout.getScene().getWindow()).close();
+					// id,비번 초기화
+				});
 				btnUpload.setOnAction(e -> upLoadDiary(e));
+				hyLinkMyPage.setOnMouseClicked(event -> {
+					messagePopup(3, "서비스 준비중입니다.");
+				});
 			}
 		});
 	}
@@ -85,13 +96,12 @@ public class MainController implements Initializable {
 			addStage.show();
 			tableViewUserList = (TableView<Diary>) parent.lookup("#tableViewUserList");
 			ConnectionDAO cDAO = new ConnectionDAO();
-			ObservableList<Diary> userDiaryList = (ObservableList<Diary>) cDAO.getUserDiaryList(userInfo.getUserid());
+			ObservableList<Diary> userDiaryList = (ObservableList<Diary>) cDAO.getUserDiaryList(userInfo.getUserid(), tableViewPreNextNum);
 			TableColumn<Diary, String> tcTitle = (TableColumn<Diary, String>) tableViewUserList.getColumns().get(0);
 			tcTitle.setCellValueFactory(new PropertyValueFactory<Diary, String>("title"));
 			TableColumn<Diary, String> tcListDate = (TableColumn<Diary, String>) tableViewUserList.getColumns().get(1);
 			tcListDate.setCellValueFactory(new PropertyValueFactory<Diary, String>("listDate"));
 			tableViewUserList.setItems(userDiaryList);
-
 			tableViewUserList.setOnMouseClicked(new EventHandler<MouseEvent>() {
 
 				@Override
@@ -102,6 +112,7 @@ public class MainController implements Initializable {
 						txtAreaContent.setText(diary.getContent());
 						stop = true;
 						labDate.setText(diary.getListDate());
+						updateDiary = diary;
 						addStage.close();
 					} else if (event.getButton() == MouseButton.SECONDARY) {
 						contextMenu(event);
@@ -110,6 +121,44 @@ public class MainController implements Initializable {
 
 				}
 			});
+			Hyperlink hyLinkPre = (Hyperlink) parent.lookup("#hyLinkPre");
+			hyLinkPre.setOnMouseClicked(new EventHandler<MouseEvent>() {
+
+				@Override
+				public void handle(MouseEvent event) {
+					System.out.println(listPageNum + " " + tableViewPreNextNum);
+					if(listPageNum >0) {
+					listPageNum--;
+					tableViewPreNextNum = tableViewPreNextNum -15;
+					ConnectionDAO cDAO = new ConnectionDAO();
+					ObservableList<Diary> userDiaryList = (ObservableList<Diary>) cDAO.getUserDiaryList(userInfo.getUserid(), tableViewPreNextNum);
+					TableColumn<Diary, String> tcTitle = (TableColumn<Diary, String>) tableViewUserList.getColumns().get(0);
+					tcTitle.setCellValueFactory(new PropertyValueFactory<Diary, String>("title"));
+					TableColumn<Diary, String> tcListDate = (TableColumn<Diary, String>) tableViewUserList.getColumns().get(1);
+					tcListDate.setCellValueFactory(new PropertyValueFactory<Diary, String>("listDate"));
+					tableViewUserList.setItems(userDiaryList);	
+					}
+				}
+			});
+			Hyperlink hyLinkNext = (Hyperlink) parent.lookup("#hyLinkNext");
+			hyLinkNext.setOnMouseClicked(new EventHandler<MouseEvent>() {
+				@Override
+				public void handle(MouseEvent event) {
+					if(listPageNum < userListCount/16) {
+						System.out.println(listPageNum + " " + tableViewPreNextNum);
+						tableViewPreNextNum = tableViewPreNextNum +15;
+						listPageNum++;
+					ConnectionDAO cDAO = new ConnectionDAO();
+					ObservableList<Diary> userDiaryList = (ObservableList<Diary>) cDAO.getUserDiaryList(userInfo.getUserid(), tableViewPreNextNum);
+					TableColumn<Diary, String> tcTitle = (TableColumn<Diary, String>) tableViewUserList.getColumns().get(0);
+					tcTitle.setCellValueFactory(new PropertyValueFactory<Diary, String>("title"));
+					TableColumn<Diary, String> tcListDate = (TableColumn<Diary, String>) tableViewUserList.getColumns().get(1);
+					tcListDate.setCellValueFactory(new PropertyValueFactory<Diary, String>("listDate"));
+					tableViewUserList.setItems(userDiaryList);
+					}
+				}
+			});
+			
 
 			Hyperlink hyLinkReturn = (Hyperlink) parent.lookup("#hyLinkReturn");
 			hyLinkReturn.setOnMouseClicked(event -> addStage.close());
@@ -125,17 +174,36 @@ public class MainController implements Initializable {
 		if (title == null || content == null || title.equals("") || content.equals("")) {
 			messagePopup(4, "일기를 작성하고 업로드를 해주세요");
 		} else {
-			Diary diary = new Diary(-1, title, content, labDate.getText());
-			ConnectionDAO cDAO = new ConnectionDAO();
-			try {
-				cDAO.insertUserDiary(userInfo.getUserid(), diary);
-				messagePopup(3, "업로드 완료");
-				changeDate();
-				txtFieldTitle.setText(null);
-				txtAreaContent.setText(null);
-			} catch (Exception e2) {
-				messagePopup(1, "업로드 실패 관리자에게 문의하세요.");
+			if (updateDiary == null) {
+				Diary diary = new Diary(-1, title, content, labDate.getText());
+				ConnectionDAO cDAO = new ConnectionDAO();
+				try {
+					cDAO.insertUserDiary(userInfo.getUserid(), diary);
+					messagePopup(3, "업로드 완료");
+					changeDate();
+					userListCount++;
+					txtFieldTitle.setText(null);
+					txtAreaContent.setText(null);
+				} catch (Exception e2) {
+					messagePopup(1, "업로드 실패 관리자에게 문의하세요.");
+				}
+
+			} else {
+				try {
+					updateDiary.setTitle(title);
+					updateDiary.setContent(content);
+					ConnectionDAO cDAO = new ConnectionDAO();
+					cDAO.updateUserDiary(userInfo.getUserid(), updateDiary);
+					messagePopup(3, "업로드 완료");
+					changeDate();
+					txtFieldTitle.setText(null);
+					txtAreaContent.setText(null);
+					updateDiary = null;
+				} catch (Exception e2) {
+					messagePopup(1, "업로드 실패 관리자에게 문의하세요.");
+				}
 			}
+
 		}
 	}
 
@@ -198,26 +266,44 @@ public class MainController implements Initializable {
 
 			@Override
 			public void handle(ActionEvent event) {
-//	            	addStage.close();
+				if (tableViewUserList.getSelectionModel().selectedItemProperty().getValue() == null) {
+					messagePopup(1, "수정할 일기가 선택되지 않았습니다.");
+				} else {
+					Diary diary = tableViewUserList.getSelectionModel().selectedItemProperty().getValue();
+					txtFieldTitle.setText(diary.getTitle());
+					txtAreaContent.setText(diary.getContent());
+					stop = true;
+					labDate.setText(diary.getListDate());
+					updateDiary = diary;
+					Stage listStage = (Stage) tableViewUserList.getScene().getWindow();
+					listStage.close();
+				}
+
 			}
 		});
 		MenuItem item2 = new MenuItem("삭제");
 		item2.setOnAction(new EventHandler<ActionEvent>() {
 			@Override
 			public void handle(ActionEvent event) {
-				ConnectionDAO cDAO = new ConnectionDAO();
-				cDAO.deleteUserDiary(userInfo.getUserid(),
-						tableViewUserList.getSelectionModel().selectedItemProperty().getValue().getListId());
-						tableViewUserList.setItems((ObservableList<Diary>) cDAO.getUserDiaryList(userInfo.getUserid()));;
+				if (tableViewUserList.getSelectionModel().selectedItemProperty().getValue() == null) {
+					messagePopup(1, "삭제할 일기가 선택되지 않았습니다.");
+				} else {
+					ConnectionDAO cDAO = new ConnectionDAO();
+					cDAO.deleteUserDiary(userInfo.getUserid(),
+							tableViewUserList.getSelectionModel().selectedItemProperty().getValue().getListId());
+					tableViewUserList.setItems((ObservableList<Diary>) cDAO.getUserDiaryList(userInfo.getUserid(), 1));
+					userListCount--;
+				}
 			}
 		});
 
 		MenuItem item3 = new MenuItem("뒤로가기");
 		item3.setOnAction(new EventHandler<ActionEvent>() {
-
 			@Override
 			public void handle(ActionEvent event) {
-				System.out.println("sss");
+				Stage listStage = (Stage) tableViewUserList.getScene().getWindow();
+				listStage.close();
+
 			}
 		});
